requirements.txt
python-multipart
weaviate-client
sentence-transformers
uvicorn[standard]
fastapi
pydantic
requests
numpy
mcp
ollama
pypdf
google-generativeai
